Move the JSONObject structure and its util function from dl package to jsonobject
