+++ jsonobject/jsonobject_test.go
@@ -1,4 +1,4 @@
-package dl
+package jsonobject
 
 import (
 	"testing"
